requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
openai
aiofiles
